GlobalBuild class and additional method invocation support
Now we have some of the logic necessary to actually start building code.
Right now, we have the ability to invoke methods in an arbitrarily
defined order, and supply arbitrary arguments to those methods. This
will comprise of our build process. Now we have to start fleshing out
the implementation of the build process.