CMake and Logger!
This is a rather large pull commit. I've added the CMake files to build
the Logger as well as the C++ source to build. I'm having one small bug
in that CMake will create an "out" directory (with a copy of the CMake
output) in the top level of the file structure. Once I figure how to
remove that, or accept it if I can't, then I'll replicate this process
with Promises and we'll be in business.