Interface for Internal Work Items and status enums
The barest of bones for FuturesFramework. So far we have an interface
that all items that can be executed in FuturesFramework must derive
from. Also, we have status enums for state checking and function
calling. This doesn't even generate a .lib file yet because no symbols
are exported (so far).