Executable Promise Code
This is more or less a derived class (WorkItem) that represents a Piece
of code that can be executed by an IScheduler (the implementation is
Scheduler) in FuturesFramework. A WorkItem is a state machine
representing its flow throughout execution. The idea of having this is
to remove templated code until the very last second and to provide a
base class for others to derive from to make their own WorkItems.