Promise to WorkItem interface as well as Continuation Support
Promises are essentially WorkItems. However, WorkItems have to be
general enough for ANYONE to derive, so Promise specific functionality
must be in a derived class. This is the derived class. In this class we
can start to implement more Promise specific code, most importantly with
the continuation aspect. I'm sorry theres no comments on any of this
yet, I'll get on that tomorrow. I have more to commit but I should
probably go to bed as its my last day of work tomorrow.