code refactoring
Split build process code into a more descriptive distribution and
refactored basic promise code. Note, this is still not all the promise
code I have, I will have to update more as I refactor. The idea of this
refactoring is to make Promises as small as they can be. There will be a
part two to this process but so far, minimizing files / objects and
removing unecessary methods / members.