refactored the build system.
It became apparent that every developing platform needs to have its own
build of googletest. So, I am setting up the build process to subdivide
the entire process into a list of building each project and linking it
to the next one (if necessary). This sets up the procedure as well as
has the testing framework ALMOST ready to go. It will link correctly,
just I was having issues with different builds of googletest. So, we
will build and cache googletest in the build process.