nevermind, I added googletest to the existing directory scheme
So, gtest is now incorporated into the current build scheme. I had to
modify some source files of gtest (just had to modify the paths for file
inclusion) so the validity of the framework remains unchanged. I
incorporated gtest when I saw exactly how much larger it was making our
codebase when it was unaltered.