TESTING
got it! We are using Catch as the testing framework, which is header
only (thank god) so it should minimize work needed to use AND we can
include it in our source tree!!!! I've set up the framework to the build
process to let us run unit tests! I'll modify the main projects (so far
it only works with testApp) but soon I'll remove testApp and get the
tests running for Logging and FuturesFramework. Let the dev train begin!