Concurrency and Promise code unit test change
Now we have worker threads!!! These guys will actually do the work! This
was the last really difficult thing to do for FuturesFramework....now we
just have to pack it all together so that clients have an entry point
into it, create continuations for Promises, and we're all set for the
basic functionality! MessageFramework can begin using it as a client!