Thread pool functionality
This is major step 1 towards providing a client entry point into the
Framework. Each Scheduler now has a (3 thread) thread pool of
WorkerThreads for each type of priority of job I can think of. Now, we
have to wrap schedulers into a singleton entry point and we're game for
client use.