fixed a unix issue..declared function before using it
Forgot to remove #include Promise.hpp from IEngine.hpp. This is bad,
because SimpleChainLinker.hpp is included in Promise.hpp, and uses
Async::GetStaticEngine() declared in IEngine.hpp.
So....SimpleChainLinker will be parsed BEFORE IEngine.hpp is done, so
GetStaticEngine() will be used BEFORE GetStaticEngine() is declared in
IEngine.hpp