fixed invalid pointer bug for continuations
No more pointer validation for parent promises, and worrying about
casting smart pointers. A std::function is used to bind the parent
result to the child's function, and a raw pointer is used. This is still
safe, as that pointer is guarenteed to exist, a continuation is executed
either from a user call (must have access to smart pointer of promise =
raw pointer exists) OR upon resolvement of the promise (raw pointer is
exisiting...WorkerThread has pointer to it)