adding Python Embedding to Robos
Robos node and Utilities ability to invoke a Python file from C++. The
Robos node is thread safe (according to the C/API documentation), but
the Utilities function is not.

TODO: Come up with a scheme for unpacking and packing C->Python and
Python->C values so clients don't have to. Should be fine for beta
release though.